amongus

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/commandbase/command/subsystemcommands/ExtensionCommand.java
@@ -15,7 +15,7 @@ public class ExtensionCommand extends SequentialCommandGroup {
     public ExtensionCommand(Robot robot, double x, double y) {
         super(
 
-            new IntakeCommand(robot.intake, Kinematics.forebar(x, y, robot.intake.FOREBAR_LENGTH)[0])
+            new IntakeCommand(robot.intake, (int) Kinematics.forebar(x, y, robot.intake.FOREBAR_LENGTH)[0])
         );
         this.intake = intake;
         this.points = Kinematics.forebar(x, y, intake.FOREBAR_LENGTH);
